Extract string builder
